ml_backend/diet_Plannar/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
uvicorn==0.25.0
flask-cors==4.0.0
